Add 'expected' sum type for ECAS
This captures the possibility of the expected value being a reference
to a reg (which isn't yet either exposed at the parser or handled in
the semantics).

Various parts of herdtools7 have been stubbed to understand expected,
but I'm unsure as to whether they are doing the correct things.  I've
commented most of the areas where I'm unsure with TODOs.