Add semantics for register-based cmpxchg
Building the semantics for the expected value now passes through
the results in an 'expected' sum type (hopefully the monadic structuring
is correct here), so we can bounce off it for the read/write events.

When the expected value is a register, we use nonatomic
read_reg/write_reg as normal; otherwise, we use read_mem/write_mem.

I've tested with one example and seem to get sensible results, but I'm
no C11 memory model expert and so there may be things I've missed.