Add statement-level CAS
Pulled out a lot of the CAS AST structure into its own record, lots of
churn arising from that.

Most of the statement CAS semantics is just the same as the expression
CAS semantics but with the value thrown away.  I'm not sure if the
`is_data` field should be true or false for CSem; I've set it as false
for now.